requirements.txt
uvicorn==0.15.0
fastapi==0.68.0
pydantic==1.8.2
sqlalchemy==1.4.23
psycopg2==2.9.1
python-dotenv==0.19.0
alembic==1.6.5
asyncpg==0.24.0
greenlet==1.1.1
pyjwt==2.1.0
